Move shape operations to files
