+++ dataset/include/scipp/dataset/dataset.h
@@ -400,10 +400,6 @@ public:
   Dataset &operator-=(const DatasetConstView &other);
   Dataset &operator*=(const DatasetConstView &other);
   Dataset &operator/=(const DatasetConstView &other);
-  Dataset &operator+=(const Dataset &other);
-  Dataset &operator-=(const Dataset &other);
-  Dataset &operator*=(const Dataset &other);
-  Dataset &operator/=(const Dataset &other);
 
   std::unordered_map<Dim, scipp::index> dimensions() const;
 
@@ -589,10 +585,6 @@ public:
   DatasetView operator-=(const DatasetConstView &other) const;
   DatasetView operator*=(const DatasetConstView &other) const;
   DatasetView operator/=(const DatasetConstView &other) const;
-  DatasetView operator+=(const Dataset &other) const;
-  DatasetView operator-=(const Dataset &other) const;
-  DatasetView operator*=(const Dataset &other) const;
-  DatasetView operator/=(const Dataset &other) const;
 
   DatasetView assign(const DatasetConstView &other) const;
 
@@ -815,22 +807,10 @@ SCIPP_DATASET_EXPORT DataArray astype(const DataArrayConstView &var,
 SCIPP_DATASET_EXPORT Dataset merge(const DatasetConstView &a,
                                    const DatasetConstView &b);
 
-SCIPP_DATASET_EXPORT DataArray concatenate(const DataArrayConstView &a,
-                                           const DataArrayConstView &b,
-                                           const Dim dim);
-SCIPP_DATASET_EXPORT Dataset concatenate(const DatasetConstView &a,
-                                         const DatasetConstView &b,
-                                         const Dim dim);
-
 SCIPP_DATASET_EXPORT DataArray rebin(const DataArrayConstView &a, const Dim dim,
                                      const VariableConstView &coord);
 SCIPP_DATASET_EXPORT Dataset rebin(const DatasetConstView &d, const Dim dim,
                                    const VariableConstView &coord);
-
-SCIPP_DATASET_EXPORT DataArray resize(const DataArrayConstView &a,
-                                      const Dim dim, const scipp::index size);
-SCIPP_DATASET_EXPORT Dataset resize(const DatasetConstView &d, const Dim dim,
-                                    const scipp::index size);
 
 [[nodiscard]] SCIPP_DATASET_EXPORT DataArray
 reciprocal(const DataArrayConstView &a);
