+++ dataset/test/dataset_operations_test.cpp
@@ -5,6 +5,7 @@
 
 #include "dataset_test_common.h"
 #include "scipp/dataset/dataset.h"
+#include "scipp/dataset/reduction.h"
 #include "test_macros.h"
 
 using namespace scipp;
