+++ dataset/test/unaligned_test.cpp
@@ -5,6 +5,7 @@
 #include "scipp/dataset/dataset.h"
 #include "scipp/dataset/event.h"
 #include "scipp/dataset/histogram.h"
+#include "scipp/dataset/shape.h"
 #include "scipp/dataset/unaligned.h"
 #include "scipp/variable/operations.h"
