+++ variable/include/scipp/variable/operations.h
@@ -8,4 +8,5 @@
 #include "scipp/variable/logical.h"
 #include "scipp/variable/misc_operations.h"
 #include "scipp/variable/reduction.h"
+#include "scipp/variable/shape.h"
 #include "scipp/variable/trigonometry.h"
