+++ variable/operations.cpp
@@ -7,7 +7,6 @@
 #include "scipp/core/dtype.h"
 #include "scipp/core/element/geometric_operations.h"
 #include "scipp/core/element/unary_operations.h"
-#include "scipp/variable/apply.h"
 #include "scipp/variable/except.h"
 #include "scipp/variable/misc_operations.h"
 #include "scipp/variable/transform.h"
@@ -30,71 +29,6 @@ std::vector<Variable> split(const Variable &var, const Dim dim,
     vars.emplace_back(var.slice({dim, indices[i], indices[i + 1]}));
   vars.emplace_back(var.slice({dim, indices.back(), var.dims()[dim]}));
   return vars;
-}
-
-Variable concatenate(const VariableConstView &a1, const VariableConstView &a2,
-                     const Dim dim) {
-  if (a1.dtype() != a2.dtype())
-    throw std::runtime_error(
-        "Cannot concatenate Variables: Data types do not match.");
-  if (a1.unit() != a2.unit())
-    throw std::runtime_error(
-        "Cannot concatenate Variables: Units do not match.");
-
-  const auto &dims1 = a1.dims();
-  const auto &dims2 = a2.dims();
-  // TODO Many things in this function should be refactored and moved in class
-  // Dimensions.
-  // TODO Special handling for edge variables.
-  for (const auto &dim1 : dims1.labels()) {
-    if (dim1 != dim) {
-      if (!dims2.contains(dim1))
-        throw std::runtime_error(
-            "Cannot concatenate Variables: Dimensions do not match.");
-      if (dims2[dim1] != dims1[dim1])
-        throw std::runtime_error(
-            "Cannot concatenate Variables: Dimension extents do not match.");
-    }
-  }
-  auto size1 = dims1.shape().size();
-  auto size2 = dims2.shape().size();
-  if (dims1.contains(dim))
-    size1--;
-  if (dims2.contains(dim))
-    size2--;
-  // This check covers the case of dims2 having extra dimensions not present in
-  // dims1.
-  // TODO Support broadcast of dimensions?
-  if (size1 != size2)
-    throw std::runtime_error(
-        "Cannot concatenate Variables: Dimensions do not match.");
-
-  Variable out(a1);
-  auto dims(dims1);
-  scipp::index extent1 = 1;
-  scipp::index extent2 = 1;
-  if (dims1.contains(dim))
-    extent1 += dims1[dim] - 1;
-  if (dims2.contains(dim))
-    extent2 += dims2[dim] - 1;
-  if (dims.contains(dim))
-    dims.resize(dim, extent1 + extent2);
-  else
-    dims.add(dim, extent1 + extent2);
-  out.setDims(dims);
-
-  out.data().copy(a1.data(), dim, 0, 0, extent1);
-  out.data().copy(a2.data(), dim, extent1, 0, extent2);
-
-  return out;
-}
-
-Variable permute(const Variable &var, const Dim dim,
-                 const std::vector<scipp::index> &indices) {
-  auto permuted(var);
-  for (scipp::index i = 0; i < scipp::size(indices); ++i)
-    permuted.data().copy(var.data(), dim, i, indices[i], indices[i] + 1);
-  return permuted;
 }
 
 Variable filter(const Variable &var, const Variable &filter) {
@@ -172,46 +106,6 @@ VariableView sqrt(const VariableConstView &var, const VariableView &out) {
 
 Variable dot(const VariableConstView &a, const VariableConstView &b) {
   return transform(a, b, element::dot);
-}
-
-Variable broadcast(const VariableConstView &var, const Dimensions &dims) {
-  if (var.dims().contains(dims))
-    return Variable{var};
-  auto newDims = var.dims();
-  const auto labels = dims.labels();
-  for (auto it = labels.end(); it != labels.begin();) {
-    --it;
-    const auto label = *it;
-    if (newDims.contains(label))
-      core::expect::dimensionMatches(newDims, label, dims[label]);
-    else
-      newDims.add(label, dims[label]);
-  }
-  Variable result(var);
-  result.setDims(newDims);
-  result.data().copy(var.data(), Dim::Invalid, 0, 0, 1);
-  return result;
-}
-
-void swap(Variable &var, const Dim dim, const scipp::index a,
-          const scipp::index b) {
-  const Variable tmp(var.slice({dim, a}));
-  var.slice({dim, a}).assign(var.slice({dim, b}));
-  var.slice({dim, b}).assign(tmp);
-}
-
-Variable resize(const VariableConstView &var, const Dim dim,
-                const scipp::index size) {
-  auto dims = var.dims();
-  dims.resize(dim, size);
-  return Variable(var, dims);
-}
-
-Variable reverse(Variable var, const Dim dim) {
-  const auto size = var.dims()[dim];
-  for (scipp::index i = 0; i < size / 2; ++i)
-    swap(var, dim, i, size - i - 1);
-  return var;
 }
 
 /// Return a deep copy of a Variable or of a VariableView.
